+++ HISTORY.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+* fix: better handle properties that happen to also be on `Object.prototype`
+
 3.19.0 / 2019-07-08
 ===================